refactor(native): isolate browser lifecycle tests
